Code cleanups

+++ moshi/src/main/java/com/squareup/moshi/JsonUtf8Writer.java
@@ -49,7 +49,7 @@ final class JsonUtf8Writer extends JsonWriter {
   static {
     REPLACEMENT_CHARS = new String[128];
     for (int i = 0; i <= 0x1f; i++) {
-      REPLACEMENT_CHARS[i] = String.format("\\u%04x", (int) i);
+      REPLACEMENT_CHARS[i] = String.format("\\u%04x", i);
     }
     REPLACEMENT_CHARS['"'] = "\\\"";
     REPLACEMENT_CHARS['\\'] = "\\\\";
